+++ src/painter/PainterRect.cpp
@@ -18,12 +18,13 @@
  */
 
 #include "PainterRect.h"
+#include <cmath>
 
 namespace
 {
-qreal smallesLenght(double lenght1, double lenght2)
+float smallesLenght(float lenght1, float lenght2)
 {
-    qreal l = (std::abs(lenght1) < std::abs(lenght2)) ? std::abs(lenght1) : std::abs(lenght2);
+    float l = (std::abs(lenght1) < std::abs(lenght2)) ? std::abs(lenght1) : std::abs(lenght2);
     return (lenght1 < 0) ? -l : l;
 }
